合并代码

+++ libraryms/view.py
@@ -22,6 +22,8 @@ import json
         GET /books
             /books/id
 '''
+
+
 # 增 POST
 @app.route('/books', methods=['POST'])
 def add_book():
@@ -39,6 +41,7 @@ def add_book():
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.ADD_BOOK_SUCCESS.value, data=None, msg=msg).__dict__)
 
+
 # 删 DELETE
 @app.route('/books/<int:id>', methods=['DELETE'])
 def delete_book(id):
@@ -53,6 +56,7 @@ def delete_book(id):
 
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.DELETE_BOOK_SUCCESS.value, data=None, msg=msg).__dict__)
+
 
 # 改 PUT
 @app.route('/books/<int:id>', methods=['PUT'])
@@ -79,6 +83,7 @@ def update_book(id):
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.UPDATE_BOOK_SUCCESS.value, data=None, msg=msg).__dict__)
 
+
 # 查 GET
 @app.route('/books', methods=['GET'])
 def get_all_books():
@@ -92,6 +97,7 @@ def get_all_books():
 
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.GET_BOOK_SUCCESS.value, data=json_books, msg=msg).__dict__)
+
 
 @app.route('/books/<int:id>', methods=['GET'])
 def get_book(id):
@@ -120,6 +126,8 @@ def get_book(id):
             /users/id
     TODO 密码加密
 '''
+
+
 # 增 POST
 @app.route('/users', methods=['POST'])
 def add_user():
@@ -137,6 +145,7 @@ def add_user():
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.ADD_USER_SUCCESS.value, data=None, msg=msg).__dict__)
 
+
 # 删 DELETE
 @app.route('/users/<int:id>', methods=['DELETE'])
 def delete_user(id):
@@ -151,6 +160,7 @@ def delete_user(id):
 
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.DELETE_USER_SUCCESS.value, data=None, msg=msg).__dict__)
+
 
 # 改 PUT
 @app.route('/users/<int:id>', methods=['PUT'])
@@ -177,6 +187,7 @@ def update_user(id):
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.UPDATE_USER_SUCCESS.value, data=None, msg=msg).__dict__)
 
+
 # 查 GET
 @app.route('/users', methods=['GET'])
 def get_all_users():
@@ -190,6 +201,7 @@ def get_all_users():
 
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.GET_USER_SUCCESS.value, data=json_users, msg=msg).__dict__)
+
 
 @app.route('/users/<int:id>', methods=['GET'])
 def get_user(id):
@@ -234,50 +246,69 @@ def add_borrow():
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.ADD_BORROW_SUCCESS.value, data=None, msg=msg).__dict__)
 
-# 删 DELETE
+# 取消某个借书申请
 @app.route('/borrows/<int:id>', methods=['DELETE'])
-def delete_borrow(id):
-    # 通过id找到对应的Borrow对象
-    borrow = Borrow.query.get(id)
-    # 删除这个对象,并提交到数据库
-    db.session.delete(borrow)
-    db.session.commit()
+def delete_borrows(id):
+    try:
+        # 查询是否存在对应的借阅记录
+        borrow = Borrow.query.get(id)
+        if not borrow:
+            return jsonify(APIResponse(ResposeCode.DELETE_BORROW_ERR.value, data="", msg='未发现该图书').__dict__)
 
-    # 响应消息
-    msg = f"删除用户{borrow.user_name}在{borrow.borrow_date}借阅{borrow.book_name}的申请"
+        db.session.delete(borrow)  # 从数据库中删除借阅记录
+        db.session.commit()
 
-    # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
-    return jsonify(APIResponse(ResposeCode.DELETE_BORROW_SUCCESS.value, data=None, msg=msg).__dict__)
+        return jsonify(APIResponse(ResposeCode.DELETE_BORROW_SUCCESS.value, data="",
+                                   msg='取消申请成功').__dict__)
+    except Exception as e:
+        return jsonify(
+            APIResponse(ResposeCode.DELETE_BORROW_ERR.value, data="", msg='error').__dict__)
 
-# 改 PUT
+# 还特定书籍与同意借阅特定书籍
 @app.route('/borrows/<int:id>', methods=['PUT'])
 def update_borrow(id):
+    try:
+        # 查询是否存在对应的借阅记录
+        borrow = Borrow.query.get(id)
+        if not borrow:
+            return jsonify(APIResponse(ResposeCode.UPDATE_BORROW_SUCCESS.value, data="", msg='未发现记录').__dict__)
 
-    # 通过id找到对应的Borrow对象
-    borrow = Borrow.query.get(id)
+        # WARNING 你这样写代码把我的路全部堵死了
+        # 将对应的借阅记录的 is_return 字段设置为 1
+        # borrow.is_return = 1
+        # db.session.commit()
 
-    # 接受前端传来的json格式的数据
-    data = request.get_json(force=True)
+        # 接受前端传来的json格式的数据
+        data = request.get_json(force=True)
 
-    for key, value in data.items():
-        # 如果这个属性存在并且值不相等，就修改这个属性的值
-        if hasattr(borrow, key) and getattr(borrow, key) != value:
-            setattr(borrow, key, value)
+        for key, value in data.items():
+            # 如果这个属性存在并且值不相等，就修改这个属性的值
+            if hasattr(borrow, key) and getattr(borrow, key) != value:
+                setattr(borrow, key, value)
 
-    # 提交到数据库
-    db.session.commit()
+        # 提交到数据库
+        db.session.commit()
 
-    # 响应消息
-    msg = ""
-    # 通过借阅
-    if borrow.is_agree == 1:
-        msg = f"通过了用户{borrow.user_name}在{borrow.borrow_date}借阅图书{borrow.book_name}的申请"
-    # 拒绝借阅
-    if borrow.is_agree == -1:
-        msg = f"拒绝了用户{borrow.user_name}在{borrow.borrow_date}借阅图书{borrow.book_name}的申请"
+        # 响应消息
+        msg = ""
+        # 判断是处理借阅申请还是归还申请
+        if borrow.is_agree != data.is_agree:
+            if borrow.is_agree == 1:
+                msg = f"通过了用户{borrow.user_name}在{borrow.borrow_date}借阅图书{borrow.book_name}的申请"
+            # 拒绝借阅
+            if borrow.is_agree == -1:
+                msg = f"拒绝了用户{borrow.user_name}在{borrow.borrow_date}借阅图书{borrow.book_name}的申请"
 
-    # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
-    return jsonify(APIResponse(ResposeCode.UPDATE_BORROW_SUCCESS.value, data=None, msg=msg).__dict__)
+        elif borrow.is_return != data.is_return:
+            if data.is_return == 1:
+                msg = f"用户{borrow.user_name}在{borrow.borrow_date}归还了图书{borrow.book_name}"
+
+        # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
+        return jsonify(APIResponse(ResposeCode.UPDATE_BORROW_SUCCESS.value, data=None, msg=msg).__dict__)
+
+
+    except Exception as e:
+        return jsonify(APIResponse(ResposeCode.UPDATE_BORROW_SUCCESS.value, data="", msg='error').__dict__)
 
 # 查 GET
 @app.route('/borrows', methods=['GET'])
@@ -292,6 +323,7 @@ def get_all_borrows():
 
     # 将返回结果封装成APIResponse对象，然后转换成json格式返回给前端
     return jsonify(APIResponse(ResposeCode.GET_BORROW_SUCCESS.value, data=json_borrows, msg=msg).__dict__)
+
 
 @app.route('/borrows/<int:id>', methods=['GET'])
 def get_borrow(id):
@@ -339,7 +371,7 @@ def login_by_account():
 @app.route('/asks/<int:user_id>', methods=['GET'])
 def get_asks(user_id):
     # 查询所有is_agree=0且borrower_id为指定用户ID的借书请求
-    asks = UBorrow.query.filter_by(is_agree=0, lender_id=user_id).all()
+    asks = U_Borrow.query.filter_by(is_agree=0, lender_id=user_id).all()
     if asks:
         ask_list = []
         for ask in asks:
@@ -365,19 +397,19 @@ def get_asks(user_id):
 # 同意他人借阅自己的图书
 @app.route('/asksAgree/<int:id>', methods=['POST'])
 def agree_asks(id):
-    borrow_request = UBorrow.query.get_or_404(id)  # 获取特定 ID 的借书请求记录
+    borrow_request = U_Borrow.query.get_or_404(id)  # 获取特定 ID 的借书请求记录
     borrow_request.is_agree = 1  # 将 is_agree 设置为 1，表示已同意
     db.session.commit()  # 提交更改到数据库
-    return jsonify(APIResponse(ResposeCode.UPDATE_UBorrow_SUCCESS.value,data='',msg='success').__dict__)
+    return jsonify(APIResponse(ResposeCode.UPDATE_UBorrow_SUCCESS.value, data='', msg='success').__dict__)
 
 
 # 拒绝
-@app.route('/asksRefuse/<int:id>',methods=['POST'])
+@app.route('/asksRefuse/<int:id>', methods=['POST'])
 def refuse_asks(id):
-    borrow_request = UBorrow.query.get_or_404(id)  # 获取特定 ID 的借书请求记录
+    borrow_request = U_Borrow.query.get_or_404(id)  # 获取特定 ID 的借书请求记录
     borrow_request.is_agree = -1  # 将 is_agree 设置为 1，表示已同意
     db.session.commit()  # 提交更改到数据库
-    return jsonify(APIResponse(ResposeCode.UPDATE_UBorrow_SUCCESS.value,data='',msg='success').__dict__)
+    return jsonify(APIResponse(ResposeCode.UPDATE_UBorrow_SUCCESS.value, data='', msg='success').__dict__)
 
 
 # 查个人借阅信息
@@ -410,46 +442,9 @@ def get_borrows(user_id):
             }
             borrow_data.append(borrow_info)
 
-        return jsonify(APIResponse(ResposeCode.GET_Borrow_SUCCESS.value, data=borrow_data, msg='success').__dict__)
+        return jsonify(APIResponse(ResposeCode.GET_BORROW_SUCCESS.value, data=borrow_data, msg='success').__dict__)
     except Exception as e:
-        return jsonify(APIResponse(ResposeCode.GET_Borrow_ERR.value, data="", msg='error').__dict__)
-
-
-# 取消某个借书申请
-@app.route('/borrows/<int:id>', methods=['DELETE'])
-def delete_borrows(id):
-    try:
-        # 查询是否存在对应的借阅记录
-        borrow = Borrow.query.get(id)
-        if not borrow:
-            return jsonify(APIResponse(ResposeCode.DELETE_Borrow_ERR.value, data="", msg='未发现该图书').__dict__)
-
-        db.session.delete(borrow)  # 从数据库中删除借阅记录
-        db.session.commit()
-
-        return jsonify(APIResponse(ResposeCode.DELETE_Borrow_SUCCESS.value, data="",
-                                   msg='取消申请成功').__dict__)
-    except Exception as e:
-        return jsonify(
-            APIResponse(ResposeCode.DELETE_Borrow_ERR.value, data="", msg='error').__dict__)
-
-
-# 还特定书籍
-@app.route('/borrows/<int:id>', methods=['PUT'])
-def update_borrows(id):
-    try:
-        # 查询是否存在对应的借阅记录
-        borrow = Borrow.query.get(id)
-        if not borrow:
-            return jsonify(APIResponse(ResposeCode.UPDATE_Borrow_ERR.value, data="", msg='未发现记录').__dict__)
-
-        # 将对应的借阅记录的 is_return 字段设置为 1
-        borrow.is_return = 1
-        db.session.commit()
-
-        return jsonify(APIResponse(ResposeCode.UPDATE_Borrow_SUCCESS.value, data="", msg='还书成功！').__dict__)
-    except Exception as e:
-        return jsonify(APIResponse(ResposeCode.UPDATE_Borrow_ERR.value, data="", msg='error').__dict__)
+        return jsonify(APIResponse(ResposeCode.GET_BORROW_ERR.value, data="", msg='error').__dict__)
 
 
 # 查借书记录
@@ -477,9 +472,9 @@ def get_borrows_history(user_id):
                 }
                 borrow_data.append(book_info)
 
-        return jsonify(APIResponse(ResposeCode.GET_Borrow_SUCCESS.value, data=borrow_data, msg='success').__dict__)
+        return jsonify(APIResponse(ResposeCode.GET_BORROW_SUCCESS.value, data=borrow_data, msg='success').__dict__)
     except Exception as e:
-        return jsonify(APIResponse(ResposeCode.GET_Borrow_ERR.value, data="", msg='error').__dict__)
+        return jsonify(APIResponse(ResposeCode.GET_BORROW_SUCCESS.value, data="", msg='error').__dict__)
 
 
 # 删除借书记录
@@ -489,17 +484,17 @@ def delete_borrows_history(id):
         # 查询是否存在对应的借阅记录
         borrow = Borrow.query.get(id)
         if not borrow:
-            return jsonify(APIResponse(ResposeCode.DELETE_Borrow_ERR.value, data="", msg='未发现记录').__dict__)
+            return jsonify(APIResponse(ResposeCode.DELETE_BORROW_SUCCESS.value, data="", msg='未发现记录').__dict__)
 
         # 删除记录
         db.session.delete(borrow)
         db.session.commit()
 
-        return jsonify(APIResponse(ResposeCode.DELETE_Borrow_SUCCESS.value, data="",
+        return jsonify(APIResponse(ResposeCode.DELETE_BORROW_SUCCESS.value, data="",
                                    msg='删除成功！').__dict__)
     except Exception as e:
         return jsonify(
-            APIResponse(ResposeCode.DELETE_Borrow_ERR.value, data="", msg='error').__dict__)
+            APIResponse(ResposeCode.DELETE_BORROW_ERR.value, data="", msg='error').__dict__)
 
 
 # 个人图书馆资源查询
@@ -507,12 +502,12 @@ def delete_borrows_history(id):
 def get_my_resources(user_id):
     try:
         # 查询属于个人图书馆的图书
-        personal_books = ULibrary.query.filter_by(user_id=user_id).all()
+        personal_books = U_Library.query.filter_by(user_id=user_id).all()
         # 封装数据
         books_data = []
         for book in personal_books:
             # 查询该图书的所有借阅记录
-            borrow_records = UBorrow.query.filter_by(book_id=book.id).all()
+            borrow_records = U_Borrow.query.filter_by(book_id=book.id).all()
             # 判断是否有符合条件的借阅记录
             status = 0
             for record in borrow_records:
@@ -540,7 +535,7 @@ def get_my_resources(user_id):
 def delete_my_resources(id):
     try:
         # 查询要删除的图书
-        book_to_delete = ULibrary.query.get(id)
+        book_to_delete = U_Library.query.get(id)
         # 如果图书不存在，返回404
         if not book_to_delete:
             return jsonify(APIResponse(ResposeCode.DELETE_UBorrow_ERR.value, data="", msg='未发现书籍！').__dict__)
@@ -559,15 +554,15 @@ def delete_my_resources(id):
 def get_other_resource(borrower_id):
     try:
         # 查询符合条件的借书信息,未审查的或者是未归还的
-        borrow_info = UBorrow.query.filter(
-            UBorrow.borrower_id == borrower_id,
-            (UBorrow.is_agree == 0) | ((UBorrow.is_agree == 1) & (UBorrow.is_return == 0))
+        borrow_info = U_Borrow.query.filter(
+            U_Borrow.borrower_id == borrower_id,
+            (U_Borrow.is_agree == 0) | ((U_Borrow.is_agree == 1) & (U_Borrow.is_return == 0))
         ).with_entities(
-            UBorrow.id,
-            UBorrow.book_id,
-            UBorrow.book_name,
-            UBorrow.borrow_date,
-            UBorrow.is_agree
+            U_Borrow.id,
+            U_Borrow.book_id,
+            U_Borrow.book_name,
+            U_Borrow.borrow_date,
+            U_Borrow.is_agree
         ).all()
 
         # 封装数据
@@ -592,7 +587,7 @@ def get_other_resource(borrower_id):
 def update_other_resources(id):
     try:
         # 查询指定ID的借书记录
-        borrow_record = UBorrow.query.get(id)
+        borrow_record = U_Borrow.query.get(id)
         if borrow_record:
             # 将借书记录的is_return字段置为1
             borrow_record.is_return = 1
@@ -609,7 +604,7 @@ def update_other_resources(id):
 def delete_other_resources(id):
     try:
         # 查询指定ID的借书记录，需要是is_agree=0还未通过的
-        borrow_record = UBorrow.query.filter_by(id=id, is_agree=0).first()
+        borrow_record = U_Borrow.query.filter_by(id=id, is_agree=0).first()
 
         if borrow_record:
             db.session.delete(borrow_record)  # 取消申请
@@ -628,7 +623,7 @@ def post_my_resources():
         # 从请求中获取书籍信息数据
         data = request.get_json()
 
-        new_book = ULibrary(
+        new_book = U_Library(
             user_id=data['user_id'],
             book_name=data['book_name'],
             author=data['author'],
@@ -661,9 +656,10 @@ def get_announcements():
             }
             announcement_data.append(announcement_info)
 
-        return jsonify(APIResponse(ResposeCode.GET_Announcement_SUCCESS.value, data=announcement_data, msg='success').__dict__)
+        return jsonify(
+            APIResponse(ResposeCode.GET_ANNOUNCEMENT_SUCCESS.value, data=announcement_data, msg='success').__dict__)
     except Exception as e:
-        return jsonify(APIResponse(ResposeCode.GET_Announcement_ERR.value, data="", msg='error').__dict__)
+        return jsonify(APIResponse(ResposeCode.GET_ANNOUNCEMENT_ERR.value, data="", msg='error').__dict__)
 
 
 # 读者咨询
@@ -684,10 +680,9 @@ def post_consults():
         db.session.add(new_consult)
         db.session.commit()
 
-        return jsonify(APIResponse(ResposeCode.ADD_Consult_SUCCESS.value, data="", msg='success').__dict__)
+        return jsonify(APIResponse(ResposeCode.ADD_CONSULT_ERR.value, data="", msg='success').__dict__)
     except Exception as e:
-        return jsonify(APIResponse(ResposeCode.ADD_Consult_SUCCESS.value, data="", msg='error').__dict__)
-
+        return jsonify(APIResponse(ResposeCode.ADD_CONSULT_SUCCESS.value, data="", msg='error').__dict__)
 
 # ——————————————————————————————————————————————————————————————————
